requirements.txt
# Version corrigées basées sur l'état actuel (Jan 2024)
streamlit==1.37.0
anthropic==0.71.0
oracledb==3.4.0
python-dotenv==1.1.1
pydantic==2.12.3
httpcore==1.0.9
httpx==0.28.1
pandas~=2.3.3